Core: Update to 10.2.7
Port From (https://github.com/TrinityCore/TrinityCore/commit/468a06baf046b716305ee665e3cfd93c80e99457)

+++ Source/Game/DataStorage/Structs/A_Records.cs
@@ -15,7 +15,7 @@ namespace Game.DataStorage
         public uint Id;
         public short InstanceID;
         public AchievementFaction Faction;
-        public ushort Supercedes;
+        public int Supercedes;
         public ushort Category;
         public byte MinimumCriteria;
         public byte Points;
